Consolidate Line and LineBlock
