+++ src/components/ComponentFactory.js
@@ -35,8 +35,8 @@ import Image from './Image';
 import Include from './Include';
 import Introduction from './Introduction';
 import Kicker from './Kicker';
-import Line from './Line';
-import LineBlock from './LineBlock';
+import Line from './LineBlock/Line';
+import LineBlock from './LineBlock/LineBlock';
 import List from './List/List';
 import ListItem from './List/ListItem';
 import ListTable from './ListTable';
