+++ src/components/LineBlock/LineBlock.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import ComponentFactory from './ComponentFactory';
+import ComponentFactory from '../ComponentFactory';
 
 const LineBlock = ({ nodeData: { children }, ...rest }) => (
   <div className="line-block">
